+++ src/Editor/index.js
@@ -30,6 +30,7 @@ import { handlePastedText } from '../utils/handlePaste';
 import Controls from '../controls';
 import getLinkDecorator from '../decorators/Link';
 import getMentionDecorators from '../decorators/Mention';
+import getQuickResponseDecorators from '../decorators/QuickResponses';
 import getHashtagDecorator from '../decorators/HashTag';
 import getBlockRenderFunc from '../renderer';
 import defaultToolbar from '../config/defaultToolbar';
@@ -67,6 +68,7 @@ export default class WysiwygEditor extends Component {
     onBlur: PropTypes.func,
     onTab: PropTypes.func,
     mention: PropTypes.object,
+    quickResponse: PropTypes.object,
     hashtag: PropTypes.object,
     textAlignment: PropTypes.string, // eslint-disable-line react/no-unused-prop-types
     readOnly: PropTypes.bool,
@@ -252,6 +254,16 @@ export default class WysiwygEditor extends Component {
         modalHandler: this.modalHandler,
       }));
     }
+    if (this.props.quickResponse) {
+      decorators.push(...getQuickResponseDecorators({
+        ...this.props.quickResponse,
+        onChange: this.onChange,
+        getEditorState: this.getEditorState,
+        getSuggestions: this.getQuickResponseSuggestions,
+        getWrapperRef: this.getWrapperRef,
+        modalHandler: this.modalHandler,
+      }));
+    }
     if (this.props.hashtag) {
       decorators.push(getHashtagDecorator(this.props.hashtag));
     }
@@ -263,6 +275,9 @@ export default class WysiwygEditor extends Component {
   getEditorState = () => this.state.editorState;
 
   getSuggestions = () => this.props.mention && this.props.mention.suggestions;
+
+  getQuickResponseSuggestions = () => this.props.quickResponse &&
+    this.props.quickResponse.suggestions;
 
   afterChange: Function = (editorState): void => {
     setTimeout(() => {
@@ -319,7 +334,7 @@ export default class WysiwygEditor extends Component {
     'defaultContentState', 'contentState', 'editorState', 'defaultEditorState', 'locale',
     'localization', 'toolbarOnFocus', 'toolbar', 'toolbarCustomButtons', 'toolbarClassName',
     'editorClassName', 'toolbarHidden', 'wrapperClassName', 'toolbarStyle', 'editorStyle',
-    'wrapperStyle', 'uploadCallback', 'onFocus', 'onBlur', 'onTab', 'mention', 'hashtag',
+    'wrapperStyle', 'uploadCallback', 'onFocus', 'onBlur', 'onTab', 'mention', 'quickResponse', 'hashtag',
     'ariaLabel', 'customBlockRenderFunc', 'customDecorators',
   ]);
 
@@ -362,7 +377,7 @@ export default class WysiwygEditor extends Component {
   };
 
   handlePastedText = (text, html) => {
-    const { editorState } = this.state;    
+    const { editorState } = this.state;
     return handlePastedText(text, html, editorState, this.onChange);
   }